Moving API class under Base
This will allow us time to reorganize the code the way we want to
